+++ tests/testthat/test-mcmc-combo.R
@@ -7,6 +7,7 @@ test_that("mcmc_combo returns a gtable object", {
   expect_gtable(mcmc_combo(arr, regex_pars = "beta"))
   expect_gtable(mcmc_combo(arr, regex_pars = "beta",
                            gg_theme = ggplot2::theme_dark()))
+  expect_gtable(mcmc_combo(drawsarr, regex_pars = "theta"))
   expect_gtable(mcmc_combo(mat, regex_pars = "beta",
                            binwidth = 1/20, combo = c("dens", "hist"),
                            facet_args = list(nrow = 2)))
@@ -18,6 +19,7 @@ test_that("mcmc_combo returns a gtable object", {
                            combo = c("trace", "hist")))
 
   expect_gtable(mcmc_combo(arr1, pars = "(Intercept)"))
+  expect_gtable(mcmc_combo(drawsarr1))
   expect_gtable(mcmc_combo(mat1))
   expect_gtable(mcmc_combo(dframe1))
 })
@@ -26,6 +28,9 @@ test_that("mcmc_combo returns a gtable object", {
 test_that("mcmc_combo throws error if 1 chain but multiple chains required", {
   expect_error(mcmc_combo(arr1chain, regex_pars = "beta",
                combo = c("trace_highlight", "dens")),
+               "requires multiple chains")
+  expect_error(mcmc_combo(drawsarr1chain, regex_pars = "theta",
+                          combo = c("trace_highlight", "dens")),
                "requires multiple chains")
   expect_error(mcmc_combo(mat, regex_pars = "beta",
                combo = c("trace_highlight", "hist")),