+++ NEWS.md
@@ -1,12 +1,23 @@
 <!-- See http://style.tidyverse.org/news.html for advice on writing news -->
 
-# bayesplot 1.8.0.9000
+# bayesplot 1.8.1.9000
 
-### New features
+* `mcmc_*()` functions now support all draws formats from the **posterior** package. (#277, @Ozan147)
+
+* `mcmc_dens()` and `mcmc_dens_overlay()` gain arguments for controlling the
+  the density calculation. (#258)
+
+* `mcmc_hist()` and `mcmc_dens()` gain argument `alpha` for controlling transparency. (#244)
+
+* `mcmc_areas()` and `mcmc_areas_ridges()` gain an argument `border_size` for
+controlling the thickness of the ridgelines. (#224)
 
 * New plotting function `ppc_km_overlay_grouped()`, the grouped variant of
-  `ppc_km_overlay()`. (#260)
+  `ppc_km_overlay()`. (#260, @fweber144)
 
+# bayesplot 1.8.1
+
+* Fix R cmd check error on linux for CRAN
 
 # bayesplot 1.8.0
 
@@ -27,7 +38,7 @@
 * New plotting function `ppc_km_overlay()` for outcome variables that are   
   right-censored. Empirical CCDF estimates of `yrep` are compared with the
   Kaplan-Meier estimate of `y`. (#233, #234, @fweber144)
-  
+
 * `ppc_loo_pit_overlay()` now uses a boundary correction for an improved kernel
   density estimation. The new argument `boundary_correction` defaults to TRUE but
   can be set to FALSE to recover the old version of the plot. (#171, #235,
@@ -36,7 +47,7 @@
 * CmdStanMCMC objects (from CmdStanR) can now be used with extractor
   functions `nuts_params()`, `log_posterior()`, `rhat()`, and
   `neff_ratio()`. (#227)
-  
+
 * On the y axis, `ppc_loo_pit_qq(..., compare = "normal")` now plots standard
   normal quantiles calculated from the PIT values (instead of the standardized
   PIT values). (#240, #243, @fweber144)